feat: rename Entity\LocalizedInterface
